move check_sensors() and parse_bytearray() into new common.js
... and export both. Import ckeck_sensors() both
from extension.js and prefs.js.  Since settings
{fan,thermal}-sensor-file now hold the **label**
of a sensor, the result of check_sensors() is required to
maps the label to a file path.
